setup testing environment
