+++ api/controllers/errors.py
@@ -1,0 +1,37 @@
+from api import app
+from flask import jsonify
+
+
+'''
+Example error handling for unprocessable entity
+'''
+@app.errorhandler(422)
+def unprocessable(error):
+    return jsonify({
+        "success": False,
+        "error": 422,
+        "message": "unprocessable"
+    }), 422
+
+
+'''
+@TODO implement error handlers using the @app.errorhandler(error) decorator
+    each error handler should return (with approprate messages):
+             jsonify({
+                    "success": False, 
+                    "error": 404,
+                    "message": "resource not found"
+                    }), 404
+
+'''
+
+'''
+@TODO implement error handler for 404
+    error handler should conform to general task above 
+'''
+
+
+'''
+@TODO implement error handler for AuthError
+    error handler should conform to general task above 
+'''
