+++ api/database/models.py
@@ -5,7 +5,8 @@ import json
 
 database_filename = "database.db"
 project_dir = os.path.dirname(os.path.abspath(__file__))
-database_path = "sqlite:///{}".format(os.path.join(project_dir, database_filename))
+database_path = "sqlite:///{}".format(os.path.join(
+    project_dir, database_filename))
 
 db = SQLAlchemy()
 
@@ -13,42 +14,54 @@ db = SQLAlchemy()
 setup_db(app)
     binds a flask application and a SQLAlchemy service
 '''
+
+
 def setup_db(app):
     app.config["SQLALCHEMY_DATABASE_URI"] = database_path
     app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
     db.app = app
     db.init_app(app)
 
+
 '''
 db_drop_and_create_all()
     drops the database tables and starts fresh
     can be used to initialize a clean database
-    !!NOTE you can change the database_filename variable to have multiple verisons of a database
+    !!NOTE you can change the database_filename variable to have multiple
+    verisons of a database
 '''
+
+
 def db_drop_and_create_all():
     db.drop_all()
     db.create_all()
+
 
 '''
 Drink
 a persistent drink entity, extends the base SQLAlchemy Model
 '''
+
+
 class Drink(db.Model):
     # Autoincrementing, unique primary key
     id = Column(Integer().with_variant(Integer, "sqlite"), primary_key=True)
     # String Title
     title = Column(String(80), unique=True)
     # the ingredients blob - this stores a lazy json blob
-    # the required datatype is [{'color': string, 'name':string, 'parts':number}]
-    recipe =  Column(String(180), nullable=False)
+    # the required datatype is [{'color': string,
+    # 'name':string, 'parts':number}]
+    recipe = Column(String(180), nullable=False)
 
     '''
     short()
         short form representation of the Drink model
     '''
+
     def short(self):
         print(json.loads(self.recipe))
-        short_recipe = [{'color': r['color'], 'parts': r['parts']} for r in json.loads(self.recipe)]
+        short_recipe = [{'color': r['color'], 'parts': r['parts']}
+                        for r in json.loads(self.recipe)]
         return {
             'id': self.id,
             'title': self.title,
@@ -59,6 +72,7 @@ class Drink(db.Model):
     long()
         long form representation of the Drink model
     '''
+
     def long(self):
         return {
             'id': self.id,
@@ -75,6 +89,7 @@ class Drink(db.Model):
             drink = Drink(title=req_title, recipe=req_recipe)
             drink.insert()
     '''
+
     def insert(self):
         db.session.add(self)
         db.session.commit()
@@ -87,6 +102,7 @@ class Drink(db.Model):
             drink = Drink(title=req_title, recipe=req_recipe)
             drink.delete()
     '''
+
     def delete(self):
         db.session.delete(self)
         db.session.commit()
@@ -100,6 +116,7 @@ class Drink(db.Model):
             drink.title = 'Black Coffee'
             drink.update()
     '''
+
     def update(self):
         db.session.commit()
