+++ requirements.txt
@@ -1,21 +1,38 @@
 astroid==2.2.5
+atomicwrites==1.3.0
+attrs==19.3.0
+autopep8==1.4.4
 Click==7.0
+colorama==0.4.1
 ecdsa==0.13.2
+entrypoints==0.3
+flake8==3.7.9
 Flask==1.0.2
+Flask-Cors==3.0.8
 Flask-SQLAlchemy==2.4.0
 future==0.17.1
+importlib-metadata==0.23
 isort==4.3.18
 itsdangerous==1.1.0
 Jinja2==2.10.1
 lazy-object-proxy==1.4.0
 MarkupSafe==1.1.1
 mccabe==0.6.1
-pycryptodome==3.3.1
-pylint==2.3.1
-python-jose-cryptodome==1.3.2
+more-itertools==7.2.0
+packaging==19.2
+pluggy==0.13.1
+py==1.8.0
+pyasn1==0.4.8
+pycodestyle==2.5.0
+pyflakes==2.1.1
+pyparsing==2.4.5
+pytest==5.3.1
+python-jose==3.0.1
+rsa==4.0
 six==1.12.0
 SQLAlchemy==1.3.3
 typed-ast==1.3.5
+wcwidth==0.1.7
 Werkzeug==0.15.2
 wrapt==1.11.1
-Flask-Cors==3.0.8
+zipp==0.6.0
